+++ llmbrix/gpt_response.py
@@ -1,6 +1,6 @@
-from typing import TypeVar
+from typing import Optional, TypeVar
 
-from pydantic import BaseModel
+from pydantic import BaseModel, model_validator
 
 from llmbrix.msg import AssistantMsg, ToolRequestMsg
 
@@ -10,8 +10,18 @@ T = TypeVar("T", bound=BaseModel)
 class GptResponse(BaseModel):
     """
     Response from a GPT model.
-    Contains assistant message and list of tool calls (potentially empty list).
+    Can contain assistant message and list of tool calls.
+    Behavior of responses API:
+     - if no tool calls are required message can be None
+     - if there are tool calls message is usually None but its not guaranteed, sometimes message can be supplied
+     - there should never be a case when message and tool calls are both None
     """
 
-    message: AssistantMsg  # response message from LLM
-    tool_calls: list[ToolRequestMsg]  # list of requested tool calls, can be and empty list
+    message: Optional[AssistantMsg] = None
+    tool_calls: Optional[list[ToolRequestMsg]] = None
+
+    @model_validator(mode="after")
+    def check_message_or_tool_calls(cls, values):
+        if values.message is None and not values.tool_calls:
+            raise ValueError("Either 'message' or 'tool_calls' must be set.")
+        return values